tools/custom-organization-policy-library/scripts/requirements.txt
PyYAML==6.0.2
yamllint==1.37.1
yapf==0.43.0